+++ examples/bounding_volume.rs
@@ -18,10 +18,10 @@ use bevy_mod_raycast::{
 fn main() {
     App::new()
         .add_plugins(DefaultPlugins.set(WindowPlugin {
-            window: WindowDescriptor {
+            primary_window: Some(Window {
                 present_mode: PresentMode::AutoNoVsync,
                 ..default()
-            },
+            }),
             ..default()
         }))
         .add_plugin(FrameTimeDiagnosticsPlugin::default())
@@ -29,21 +29,23 @@ fn main() {
         // You will need to pay attention to what order you add systems! Putting them in the wrong
         // order can result in multiple frames of latency. Ray casting should probably happen after
         // the positions of your meshes have been updated in the UPDATE stage.
-        .add_system_to_stage(
-            CoreStage::First,
-            update_raycast_with_cursor.before(RaycastSystem::BuildRays::<MyRaycastSet>),
+        .add_system(
+            update_raycast_with_cursor
+                .before(RaycastSystem::BuildRays::<MyRaycastSet>)
+                .in_base_set(CoreSet::First),
         )
         .add_startup_system(setup_scene)
         .add_startup_system(setup_ui)
         .add_system(update_fps)
         .add_system(make_scene_pickable)
-        .add_system_to_stage(CoreStage::First, manage_aabb)
+        .add_system(manage_aabb.in_base_set(CoreSet::First))
         .run();
 }
 
 // This is a unit struct we will use to mark our generic `RaycastMesh`s and `RaycastSource` as part
 // of the same group, or "RaycastSet". For more complex use cases, you might use this to associate
 // some meshes with one ray casting source, and other meshes with a different ray casting source."
+#[derive(Clone, Reflect)]
 struct MyRaycastSet;
 
 // Update our `RaycastSource` with the current cursor position every frame.