+++ examples/ray_intersection_over_mesh.rs
@@ -16,18 +16,19 @@ use bevy_mod_raycast::{
 fn main() {
     App::new()
         .add_plugins(DefaultPlugins.set(WindowPlugin {
-            window: WindowDescriptor {
+            primary_window: Some(Window {
                 present_mode: PresentMode::AutoNoVsync, // Reduces input latency
                 ..default()
-            },
+            }),
             ..default()
         }))
         .add_plugin(DefaultRaycastingPlugin::<Ground>::default())
         .add_startup_system(setup)
         .add_startup_system(setup_ui)
-        .add_system_to_stage(
-            CoreStage::First,
-            update_raycast_with_cursor.before(RaycastSystem::BuildRays::<Ground>),
+        .add_system(
+            update_raycast_with_cursor
+                .before(RaycastSystem::BuildRays::<Ground>)
+                .in_base_set(CoreSet::First),
         )
         .add_system(check_path)
         .add_system(move_origin)
@@ -85,7 +86,7 @@ fn setup_ui(mut commands: Commands, asset_server: Res<AssetServer>) {
 struct PathStatus;
 
 // Marker struct for the ground, used to get cursor position
-#[derive(Component)]
+#[derive(Component, Reflect, Clone)]
 struct Ground;
 
 // Marker struct for the path origin, shown by a cyan sphere
@@ -122,7 +123,7 @@ fn setup(
     // Spawn a plane that will represent the ground. It will be used to pick the mouse location in 3D space
     commands
         .spawn(PbrBundle {
-            mesh: meshes.add(Mesh::from(shape::Plane { size: 500.0 })),
+            mesh: meshes.add(Mesh::from(shape::Plane::from_size(500.0))),
             material: materials.add(Color::DARK_GRAY.into()),
             ..Default::default()
         })
@@ -171,14 +172,14 @@ fn setup(
     // Spawn the intersection point, invisible by default until there is an intersection
     commands
         .spawn(PbrBundle {
-            mesh: meshes.add(Mesh::from(shape::Icosphere::default())),
+            mesh: meshes.add(Mesh::try_from(shape::Icosphere::default()).unwrap()),
             material: materials.add(StandardMaterial {
                 unlit: true,
                 base_color: Color::RED,
                 ..Default::default()
             }),
             transform: Transform::from_scale(Vec3::splat(0.1)),
-            visibility: Visibility { is_visible: false },
+            visibility: Visibility::Hidden,
             ..Default::default()
         })
         .insert(PathObstaclePoint);
@@ -272,7 +273,7 @@ fn check_path(
                     // Set everything as OK in case there are no obstacle in path
                     text.sections[1].value = "Direct!".to_string();
                     text.sections[1].style.color = Color::GREEN;
-                    visible.is_visible = false;
+                    *visible = Visibility::Hidden;
 
                     let mut closest_hit = f32::MAX;
 
@@ -296,7 +297,7 @@ fn check_path(
                                     text.sections[1].value = "Obstructed!".to_string();
                                     text.sections[1].style.color = Color::RED;
                                     intersection_transform.translation = intersection.position();
-                                    visible.is_visible = true;
+                                    *visible = Visibility::Inherited;
                                     closest_hit = hit_distance;
                                 }
                             }